refactor: 테스트 영역 구분, 테스트 초기화 수정, 불필요한 테스트 코드 정리
- [x] 4-3.리뷰어님 피드백 반영
        - [x] 4-3-1.생성자 접근제어자 `public` -> `protected` 변경
        - [x] 4-3-2.getter/setter 제거
        - [x] 4-3-3.반복적인 필드 추상클래스 활용
            - [x] 4-3-3-1.중복된 필드 모아놓기
            - [x] 4-3-3-2.jpa auditing 적용
        - [x] 4-3-4.테스트 영역 구분
            - [x] 4-3-4-1.`AnswerTest` -> `AnswerTest`, `AnswerRepositoryTest`
            - [x] 4-3-4-2.`DeleteHistoryTest` -> `DeleteHistoryTest`, `DeleteHistoryRepositoryTest`
            - [x] 4-3-4-3.`QuestionTest` -> `QuestionTest`, `QuestionRepositoryTest`
            - [x] 4-3-4-4.`UserTest` -> `UserTest`, `UserRepositoryTest`
        - [x] 4-3-5.테스트 초기화 수정
            - [x] 4-3-5-1.`AnswerTest`
            - [x] 4-3-5-2.`AnswerRepositoryTest`
            - [x] 4-3-5-3.`DeleteHistoryTest`
            - [x] 4-3-5-4.`DeleteHistoryRepositoryTest`
            - [x] 4-3-5-5.`QuestionTest`
            - [x] 4-3-5-6.`QuestionRepositoryTest`
            - [x] 4-3-5-7.`UserTest`
            - [x] 4-3-5-8.`UserRepositoryTest`
        - [x] 4-3-6.불필요한 테스트 코드 정리
            - [x] 4-3-6-1.`AnswerTest`
            - [x] 4-3-6-2.`AnswerRepositoryTest`
            - [x] 4-3-6-3.`DeleteHistoryTest`
            - [x] 4-3-6-4.`DeleteHistoryRepositoryTest`
            - [x] 4-3-6-5.`QuestionTest`
            - [x] 4-3-6-6.`QuestionRepositoryTest`
            - [x] 4-3-6-7.`UserTest`
            - [x] 4-3-6-8.`UserRepositoryTest`

Resolve: #134
See Also:
