+++ src/main/java/qna/domain/User.java
@@ -30,6 +30,10 @@ public class User extends BaseEntity {
 	protected User() {
 	}
 
+	public User(String userId, String password, String name, String email) {
+		this(null, userId, password, name, email);
+	}
+
 	public User(Long id, String userId, String password, String name, String email) {
 		this.id = id;
 		this.userId = userId;
